Backend-API/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
asgiref==3.8.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-cors-headers
async-timeout==5.0.1
    # via redis
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==25.4.16
    # via twisted
billiard==4.2.1
    # via celery
boto3==1.38.38
    # via -r requirements.in
botocore==1.38.38
    # via
    #   boto3
    #   s3transfer
celery==5.4.0
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-celery-results
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
channels==4.2.2
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.2.1
    # via -r requirements.in
charset-normalizer==3.3.2    
# via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
constantly==23.10.4
    # via twisted
cron-descriptor==1.4.5
    # via django-celery-beat
cryptography==45.0.6
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.1.2
    # via -r requirements.in
distro==1.9.0
    # via openai
django==5.1.7
    # via
    #   -r requirements.in
    #   channels
    #   django-celery-beat
    #   django-celery-results
    #   django-cors-headers
    #   django-csp
    #   django-extensions
    #   django-filter
    #   django-redis
    #   django-simple-history
    #   django-sslserver
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-celery-beat==2.7.0
    # via -r requirements.in
django-celery-results==2.5.1
    # via -r requirements.in
django-cors-headers==4.6.0
    # via -r requirements.in
django-csp==3.8
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
django-redis==6.0.0
    # via -r requirements.in
django-simple-history==3.7.0
    # via -r requirements.in
django-sslserver==0.22
    # via -r requirements.in
django-storages==1.14.6
    # via -r requirements.in
django-timezone-field==7.0
    # via
    #   -r requirements.in
    #   django-celery-beat
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
dj-database-url==2.1.0
    # via -r requirements.in
et-xmlfile==2.0.0
    # via openpyxl
exceptiongroup==1.3.0
    # via anyio
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
    # via
    #   -r requirements.in
    #   openai
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
    #   twisted
incremental==24.7.2
    # via twisted
inflection==0.5.1
    # via drf-spectacular
jiter==0.10.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
kombu==5.4.2
    # via celery
msgpack==1.1.0
    # via channels-redis
numpy==2.2.6
    # via
    #   -r requirements.in
    #   pandas
openai==1.84.0
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
packaging==24.2
    # via gunicorn
pandas==2.3.1
    # via -r requirements.in
pillow==11.0.0
    # via -r requirements.in
prompt-toolkit==3.0.48
    # via click-repl
psycopg2-binary==2.9.10
    # via -r requirements.in

# OCR

pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via openai
pydantic-core==2.33.2
    # via pydantic
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyopenssl==25.1.0
    # via twisted
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   botocore
    #   celery
    #   pandas
    #   python-crontab
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via
    #   -r requirements.in
    #   pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   drf-spectacular
redis==5.2.1
    # via
    #   -r requirements.in
    #   channels-redis
    #   django-redis
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via -r requirements.in
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
s3transfer==0.13.0
    # via boto3
service-identity==24.2.0
python-json-logger==2.0.7
    # via twisted
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlparse==0.5.2
    # via django
tomli==2.2.1
    # via incremental
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   openai
twisted[tls]==25.5.0
    # via daphne
txaio==25.6.1
    # via autobahn
typing-extensions==4.12.2
    # via
    #   anyio
    #   asgiref
    #   exceptiongroup
    #   openai
    #   pydantic
    #   pydantic-core
    #   pyopenssl
    #   twisted
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2024.2
    # via
    #   -r requirements.in
    #   celery
    #   django-celery-beat
    #   kombu
    #   pandas
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.4.0
    # via
    #   botocore
    #   requests
uvicorn==0.30.6
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
zope-interface==7.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
